+++ src/ui/frame/mod.rs
@@ -6,20 +6,18 @@
 #![doc = crate::_doc!(hr)]
 //
 
-#[cfg(test)]
-mod _test;
-
 mod id; // UiId, UiKey, UiScope
 mod frame; // UiFrame, UiPhase
 // mod mem; // UiMemory
 // mod cache; // UiCache
-// mod output; // UiOutput
+mod output; // UiOutput
 
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             id::*,
             frame::*,
+            output::*,
         };
     }
 }